Expand estimate.py into components
