+++ .gitignore
@@ -41,3 +41,6 @@ desktop.ini
 .vscode/c_cpp_properties.json
 .vscode/launch.json
 .pio
+
+# Python
+*.pyc
